feat: implement users schema with entity, service, and comprehensive tests
- Add UserEntity JPA class with full validation annotations
- Fields: id, email, firstName, lastName, phoneNumber, createdAt, updatedAt
- JPA annotations: @Entity, @Table, @Id, @GeneratedValue, @Column with constraints
- Jakarta validations: @NotNull, @Email, @Size
- Lombok: @Data, @NoArgsConstructor, @AllArgsConstructor
- Add UserService with CRUD operations and validation logic
- Update UserRepository to support Long ID type
- Add 15 unit tests for UserEntity validation
- Add 18 service tests with Mockito for all operations
- All 34 tests passing

Acceptance Criteria Met:
✅ UserEntity.java exists with proper annotations
✅ @Entity and @Table annotations applied
✅ All required fields with correct types
✅ JPA annotations with nullable, unique, length properties
✅ Jakarta validation annotations (@NotNull, @Email, @Size)
✅ Lombok annotations for automatic getters/setters
✅ Entity compiles without errors
✅ Validation annotations implemented
✅ Service layer supports CRUD operations

+++ src/main/java/com/rocketFoodDelivery/rocketFood/repository/UserRepository.java
@@ -8,8 +8,8 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
-    Optional<UserEntity> findById(int id);
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+    Optional<UserEntity> findById(Long id);
     List<UserEntity> findAllByOrderByIdDesc();
     Optional<UserEntity> findByEmail(String email);
 }
